Update real-world example to React 0.14 and Redux DevTools Beta 3
