+++ examples/real-world/components/Explore.js
@@ -50,6 +50,9 @@ export default class Explore extends Component {
         <p>
           Code on <a href={GITHUB_REPO} target="_blank">Github</a>.
         </p>
+        <p>
+          Move the DevTools with Ctrl+W or hide them with Ctrl+H.
+        </p>
       </div>
     );
   }
