+++ examples/real-world/store/configureStore.js
@@ -1,30 +1,5 @@
-import { createStore, applyMiddleware, compose } from 'redux';
-import { reduxReactRouter } from 'redux-router';
-import { devTools } from 'redux-devtools';
-import createHistory from 'history/lib/createBrowserHistory';
-import routes from '../routes';
-import thunk from 'redux-thunk';
-import api from '../middleware/api';
-import createLogger from 'redux-logger';
-import rootReducer from '../reducers';
-
-const finalCreateStore = compose(
-  applyMiddleware(thunk, api),
-  reduxReactRouter({ routes, createHistory }),
-  applyMiddleware(createLogger()),
-  devTools()
-)(createStore);
-
-export default function configureStore(initialState) {
-  const store = finalCreateStore(rootReducer, initialState);
-
-  if (module.hot) {
-    // Enable Webpack hot module replacement for reducers
-    module.hot.accept('../reducers', () => {
-      const nextRootReducer = require('../reducers');
-      store.replaceReducer(nextRootReducer);
-    });
-  }
-
-  return store;
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./configureStore.prod');
+} else {
+  module.exports = require('./configureStore.dev');
 }
